complete review (admin &staff side)

+++ src/java/controller/EditProdServlet.java
@@ -7,6 +7,7 @@ package controller;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.math.BigDecimal;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -150,7 +151,7 @@ public class EditProdServlet extends HttpServlet {
 
             
             prod.setProductname(prodName);
-            prod.setPrice(prodPrice);
+            prod.setPrice(BigDecimal.valueOf(prodPrice));
             prod.setQuantity(prodStock);
             prod.setProductdescription(prodDesc);
             prod.setStatus(prodStatus);
